refactor: HandlerMapper가 상태를 가지도록 변경
